# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kassomat API
  version: 1.0.0
  description: |
    Kassomat POS SaaS API — RKSV-konforme Registrierkasse für österreichische Betriebe.

    ## Authentifizierung
    Alle Endpoints (außer /auth/* und /webhooks/*) erfordern einen Bearer Token im
    `Authorization` Header: `Authorization: Bearer <access_token>`

    ## Multi-Tenant
    Alle Daten sind tenant-isoliert. Der Tenant wird aus dem JWT extrahiert.
    Ein Benutzer kann niemals Daten eines anderen Tenants sehen oder verändern.

    ## Geldbeträge
    Alle Geldbeträge sind in **Cent (Integer)**. Beispiel: €2,50 = `250`
servers:
  - url: https://api.kassomat.at
    description: Production
  - url: https://api-staging.kassomat.at
    description: Staging
  - url: http://localhost:3001
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ============================================================
  # REUSABLE SCHEMAS
  # ============================================================
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Ungültiger oder abgelaufener Token
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    Tenant:
      type: object
      required: [id, name, slug, plan, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Spätii Innsbruck
        slug:
          type: string
          example: spaetii-innsbruck
        plan:
          type: string
          enum: [starter, pro, business]
        status:
          type: string
          enum: [active, suspended, trial]
        trialEndsAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        settings:
          $ref: '#/components/schemas/TenantSettings'
    TenantSettings:
      type: object
      properties:
        currency:
          type: string
          example: EUR
        timezone:
          type: string
          example: Europe/Vienna
        vatNumber:
          type: string
          nullable: true
          example: ATU12345678
        receiptFooter:
          type: string
          nullable: true
        printerIp:
          type: string
          nullable: true
          example: 192.168.1.100
        printerPort:
          type: integer
          nullable: true
          example: 9100
        rksvEnabled:
          type: boolean
        atrust:
          $ref: '#/components/schemas/ATrustConfig'
          nullable: true
        lieferando:
          $ref: '#/components/schemas/LieferandoConfig'
          nullable: true
        wix:
          $ref: '#/components/schemas/WixConfig'
          nullable: true
    ATrustConfig:
      type: object
      required: [certificateSerial, environment]
      properties:
        certificateSerial:
          type: string
        environment:
          type: string
          enum: [test, production]
    LieferandoConfig:
      type: object
      required: [restaurantId, isActive]
      properties:
        restaurantId:
          type: string
        isActive:
          type: boolean
    WixConfig:
      type: object
      required: [siteId, isActive, defaultDeliveryPayment]
      properties:
        siteId:
          type: string
        isActive:
          type: boolean
        defaultDeliveryPayment:
          type: string
          enum: [cash, online]
    User:
      type: object
      required: [id, tenantId, email, role, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [owner, admin, cashier]
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
...
    Shift:
      type: object
      required: [id, tenantId, cashierId, startedAt, openingFloat, totalRevenue, receiptCount]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        cashierId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        openingFloat:
          type: integer
        closingFloat:
          type: integer
          nullable: true
        totalRevenue:
          type: integer
        receiptCount:
          type: integer
    DailyClosing:
      type: object
      required: [id, tenantId, date, closedBy, closedAt, totalRevenue, receiptCount]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        date:
          type: string
          example: "2024-03-15"
        closedBy:
          type: string
          format: uuid
        closedAt:
          type: string
          format: date-time
        totalCash:
          type: integer
        totalCard:
          type: integer
        totalOnline:
          type: integer
        totalRevenue:
          type: integer
        receiptCount:
          type: integer
        cancellationCount:
          type: integer
        depExportPath:
          type: string
          nullable: true

  # ============================================================
  # REUSABLE RESPONSES
  # ============================================================
  responses:
    Unauthorized:
      description: Kein oder ungültiger JWT Token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Keine Berechtigung für diese Aktion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validierungsfehler in den Eingabedaten
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []

# ============================================================
# PATHS
# ============================================================
paths:
...
  /auth/login:
    post:
      summary: Login mit E-Mail und Passwort
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Login erfolgreich
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Access Token erneuern mit Refresh Token
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token erneuert
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Ausloggen und Refresh Token invalidieren
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logout erfolgreich

  # ---- TENANT ----
  /tenant:
    get:
      summary: Eigenen Tenant abrufen
      tags: [Tenant]
      responses:
        '200':
          description: Tenant-Daten
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      summary: Tenant-Einstellungen aktualisieren
      tags: [Tenant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                receiptFooter:
                  type: string
                  nullable: true
                printerIp:
                  type: string
                  nullable: true
                printerPort:
                  type: integer
                  nullable: true
                vatNumber:
                  type: string
                  nullable: true
                rksvEnabled:
                  type: boolean
                atrust:
                  type: object
                  properties:
                    certificateSerial:
                      type: string
                    apiKey:
                      type: string
                    environment:
                      type: string
                      enum: [test, production]
                lieferando:
                  type: object
                  properties:
                    restaurantId:
                      type: string
                    apiKey:
                      type: string
                    webhookSecret:
                      type: string
                    isActive:
                      type: boolean
                wix:
                  type: object
                  properties:
                    siteId:
                      type: string
                    apiKey:
                      type: string
                    webhookSecret:
                      type: string
                    isActive:
                      type: boolean
                    defaultDeliveryPayment:
                      type: string
                      enum: [cash, online]
      responses:
        '200':
          description: Einstellungen aktualisiert
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /tenant/register:
    post:
      summary: Neuen Tenant anlegen (Signup)
      security: []
      tags: [Tenant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantName, ownerEmail, ownerPassword, ownerName]
              properties:
                tenantName:
                  type: string
                  minLength: 2
                  example: Spätii Innsbruck
                ownerEmail:
                  type: string
                  format: email
                ownerPassword:
                  type: string
                  minLength: 8
                ownerName:
                  type: string
                vatNumber:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Tenant erstellt, Trial gestartet
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: E-Mail bereits registriert

  # ---- PRODUCTS ----
...
  /daily-closing:
    post:
      summary: Tagesabschluss durchführen
      tags: [Receipts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cashCount]
              properties:
                cashCount:
                  type: integer
                  description: Gezählter Kassenbestand in Cent
                notes:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Tagesabschluss abgeschlossen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DailyClosing'

  # ---- ORDERS ----
...
  /dep/export/range:
    get:
      summary: DEP-Export für einen Zeitraum
      tags: [DEP]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: DEP Export für Zeitraum

  # ---- WEBHOOKS ----
...
  /shifts:
    get:
      summary: Alle Schichten abrufen
      tags: [Shifts]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Liste der Schichten
  /shifts/start:
    post:
      summary: Neue Schicht starten
      tags: [Shifts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [openingFloat]
              properties:
                openingFloat:
                  type: integer
                  description: Kassenbestand bei Schichtstart in Cent
      responses:
        '201':
          description: Schicht gestartet
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Shift'
  /shifts/end:
    post:
      summary: Aktuelle Schicht beenden
      tags: [Shifts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [closingFloat]
              properties:
                closingFloat:
                  type: integer
                  description: Gezählter Kassenbestand in Cent
      responses:
        '200':
          description: Schicht beendet
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Shift'

  # ---- ANALYTICS ----
...
  /analytics/range:
    get:
      summary: Analytics für einen Zeitraum
      tags: [Analytics]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        '200':
          description: Analytics für Zeitraum